fixed implementation and tests
