+++ src/main/java/org/folio/innreach/controller/InnReachTransactionController.java
@@ -5,7 +5,6 @@ import java.util.stream.Collectors;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.folio.innreach.dto.InnReachRecallItemDTO;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -14,7 +13,6 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -103,12 +101,4 @@ public class InnReachTransactionController implements InnReachTransactionApi {
     return response;
   }
 
-  @Override
-  @PostMapping("/inn-reach/d2ir/circ/recall/{trackingId}/{centralCode}")
-  public ResponseEntity<InnReachResponseDTO> recallItem(@PathVariable String trackingId,
-                                                        @PathVariable String centralCode,
-                                                        InnReachRecallItemDTO recallItem) {
-    var response = transactionService.recallItem(trackingId, centralCode, recallItem);
-    return ResponseEntity.ok(response);
-  }
 }
