+++ src/main/java/org/folio/innreach/domain/service/InnReachTransactionService.java
@@ -7,7 +7,6 @@ import org.folio.innreach.dto.InnReachResponseDTO;
 import org.folio.innreach.dto.InnReachTransactionDTO;
 import org.folio.innreach.dto.InnReachTransactionFilterParametersDTO;
 import org.folio.innreach.dto.InnReachTransactionsDTO;
-import org.folio.innreach.dto.InnReachRecallItemDTO;
 import org.folio.innreach.dto.TransactionHoldDTO;
 
 public interface InnReachTransactionService {
@@ -18,6 +17,4 @@ public interface InnReachTransactionService {
   InnReachTransactionDTO getInnReachTransaction(UUID transactionId);
 
   InnReachTransactionsDTO getAllTransactions(Integer offset, Integer limit, InnReachTransactionFilterParametersDTO parameters);
-
-  InnReachResponseDTO recallItem(String trackingId, String centralCode, InnReachRecallItemDTO recallItem);
 }
